refactor: update runtime
