+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   preset: '@nuxt/test-utils',
   moduleNameMapper: {
-    '~image/(.*)': '<rootDir>/src/runtime/$1',
+    '@nuxt/image/runtime/(.*)': '<rootDir>/src/runtime/$1',
     '~/(.*)': '<rootDir>/$1',
     '^.+\\.css$': '<rootDir>/test/fixture/utils/CSSStub.js'
   }
