+++ playground/nuxt.config.ts
@@ -7,9 +7,6 @@ export default {
   buildModules: [
     '@nuxt/typescript-build'
   ],
-  plugins: [
-    'plugins/dat.gui.client.js'
-  ],
   image: {
     accept: ['nuxtjs.org', 'unsplash.com'],
     twicpics: {
