--- playground/plugins/dat.gui.client.js
@@ -1,4 +1,0 @@
-import Vue from 'vue'
-import DatGui from '@cyrilf/vue-dat-gui'
-
-Vue.use(DatGui)
