+++ src/runtime/index.ts
@@ -1,5 +1,4 @@
 export * from './image'
-export * from './types'
 export * from './utils'
 
 export { NuxtImg } from './components/nuxt-img'
